prepare some helper-functions
