+++ src/assets/js/helper/addClass.js
@@ -1,0 +1,5 @@
+function addClass(element, className) {
+	return element.classList.add(className);
+}
+
+export default addClass;
