+++ src/assets/js/helper/find.js
@@ -1,0 +1,5 @@
+function find(term) {
+	return document.querySelectorAll(term);
+}
+
+export default find;
