+++ src/assets/js/helper/removeClass.js
@@ -1,0 +1,5 @@
+function addClass(element, className) {
+	return element.classList.remove(className);
+}
+
+export default addClass;
